test(e2e): map TINYINT/SMALLINT/MEDIUMINT to int, matching the ingestor
_get_sqlalchemy_type coerces all three to SQLAlchemy Integer -> MySQL int, so information_schema reports 'int'; the dtype table wrongly expected distinct names and would falsely fail dimension (d) for any fixture declaring those types. Latent today (no fixture uses them) but corrected. Bugbot finding on #366.

+++ e2e/test_content_compare_e2e.py
@@ -91,9 +91,12 @@ SCHEMA_TO_INFORMATION_SCHEMA = {
     "TEXT": "text",
     "INT": "int",
     "INTEGER": "int",
-    "TINYINT": "tinyint",
-    "SMALLINT": "smallint",
-    "MEDIUMINT": "mediumint",
+    # TINYINT/SMALLINT/MEDIUMINT all map to SQLAlchemy Integer in
+    # database._get_sqlalchemy_type, which MySQL renders as `int` — so
+    # information_schema reports `int`, not the narrower declared width.
+    "TINYINT": "int",
+    "SMALLINT": "int",
+    "MEDIUMINT": "int",
     "BIGINT": "bigint",
     "FLOAT": "float",
     "DOUBLE": "double",
